Migrate kafka monitor to use nosetests
