+++ .gitignore
@@ -38,3 +38,14 @@ var/
 *.egg-info/
 .installed.cfg
 *.egg
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*,cover
+.hypothesis/
